Add stub of a Path enum
This will abstract over both bezier and hyperbezier paths.
